Make temporary auth.conf in acceptance test readable by Puppet
The create_remote_file method uploads files with mode 600 since that is the
default mode for Tempfile. The puppet user thus isn't able to read the
temporary auth.conf, so it just uses the default rules instead, causing this
test to fail.

Also, this test was dependent on running after another test which set up a cert
on the agents. Since that can't be guaranteed, the master now runs with
autosign to ensure the agents can connect.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>