Cleanup indentation, comment, and unused code
The Great Reindentation of '10 left certain structures indented incorrectly;
this addresses some of these instances.

The comment about loading all providers incorrectly stated that we're trying to
figure out the type, when we're actually trying to figure out the provider.

There was an unused variable initialization that was introduced in 2b14f627,
which was reverting c19835c, 9290cc8, and ffb4c2d.

Paired-with: Jacob Helwig <jacob@puppetlabs.com>