(#4416) Ensure types are providified after reloading
Previously, the 'provider' parameter for a type was only added when creating a
provider for that type. This would cause a type to forget about its 'provider'
parameter when only the type was reloaded.

This was manifesting itself in pluginsync, when a provider plugin would be
loaded before its type, causing the type to be autoloaded. The type plugin
would then be loaded again by the plugin handler. Because the type => provider
information is stored separately from the type, the providers don't need to be
reloaded, and thus don't recreate the type's 'provider' parameter.

Now we always "providify" the type (add its 'provider' parameter) upon
creation, after trying to load its providers, if any providers are present.

Paired-with: Jacob Helwig <jacob@puppetlabs.com>