(#7224) Add a helper to Puppet::SSL::Certificate to retrieve alternate names
Alternate names, if present, are specified in the subjectAltName extension of
the certificate. The values are in the form:

"DNS:alternate_name1, DNS:alternate_name2"

This helper will retrieve the value of the subjectAltName extension and extract
the alternate names, returning and empty list if the extension is absent. This
will make it easier to access the entire list of possible names for a
certificate, rather than just the common name; this is helpful for generating
more detailed SSL error messages.

Paired-With: Jacob Helwig <jacob@puppetlabs.com>

Conflicts:

	lib/puppet/ssl/certificate.rb
	spec/unit/ssl/certificate_spec.rb

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>