(#7114) Add tests for option property
The option property of the ssh_authorized_key type supports an array of
options. To prevent user errors we fail when the user tries to pass
multiple option as a single string like "option1,option2" instead of an
array. A single option however can also include a comma. Sample option:
from="host1,host2".

Add tests that demonstrate that the new value validation on options
is to strict.