Add a service provider that manages systemd services natively
systemd is the default system/service manager in Fedora 15 and will
likely become the default on other Linux distributions in the future.
While systemd provides a compatibility layer that allows the standard
"redhat" Puppet service provider to manage some services on a systemd
system, as more services are converted to native systemd services the
compatibility layer will become less and less useful.  For more
information on systemd see:

http://www.freedesktop.org/wiki/Software/systemd