(#8032) Add containment to create_resources
Without this change native resource types declared using the
create_resources() function are not contained within the class scope of
the function call.  As a result, resources were "floating off" in the
graph, disconnected from the rest of the relationship edges.

With this change, the scope is preserved and native resources are
contained by the class the function call is executed from.

Reviewed-by: Dan Bode <dan@puppetlabs.com>