(#8048) Gem install puppet no longer fails if rdoc enabled.
Pre-released versions of faces may have used the "desc" option, which
was later changed to "description", and an alias provided for "desc". The
previous fix for 8048 removed this backwards compatibility (since it was
never released to customers). This commit fixes the test case that
assumed the "desc" option was still available.

Reviewed-By: Daniel Pittman <daniel@puppetlabs.com>