(#7581) Provide more detailed error message when missing gems on Windows
Running Puppet on Windows requires the sys-admin, win32-process &
win32-dir gems.  If any of these gems were missing, Puppet would fail
with the message "Cannot determine basic system flavour".

When trying to determine if we are on Windows, we now warn with the
message "Cannot run on Microsoft Windows without the sys-admin,
win32-process & win32-dir gems: #{err}", where err is the normal ruby
load error message stating which gem could not be loaded.

We also only warn if the POSIX feature is not present.

Signed-off-by: James Turnbull <james@puppetlabs.com>
Signed-off-by: Jacob Helwig <jacob@puppetlabs.com>
Reviewed-by: Cameron Thomas <cameron@puppetlabs.com>