Fixes #8119 - Write to a temporary file and move when storing reports.
When writing reports, there is a window in between opening and writing
to the report file when the report file exists as an empty file. This
makes writing report processors a little annoying as they have to deal
with this case. This patch writes the report into a temporary file then
renames it to the report file, which eliminates this window.