Disable the master on Windows instead of blowing up with failed resources
Running the Puppet master on Windows is not supported, so instead of
failing with what can be cryptic error messages about failed resources
we fail with an explicit error message about the master on Windows not
being supported.  This way a user isn't mistakenly given the
impression that running a master on Windows will work, and they just
have something mis-configured.

Signed-off-by: Jacob Helwig <jacob@puppetlabs.com>
Reviewed-by: Max Martin <max@puppetlabs.com>