(#8418) Fix inspect app to have the correct run_mode
Requiring puppet before the run_mode has been set by the application
causes the default run_mode of 'user' to be set instead of what the
application wants.  This leads to the incorrect default settings to be
used, which lead to inspect not being able to properly retrieve file
metadata from a fileserver.

Reviewed-by: Max Martin <max@puppetlabs.com>