maint: SSL::Inventory.serial should report missing names.
Our SSL inventory was able to find the serial number of a certificate by name,
but was incapable of living up to the contract it offered, that it would
actually report when a certificate was missing.

Now it returns `nil`, which is the same case as "no inventory", if the
certificate was not found, rather than accidentally returning the entire
inventory data as raw strings.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>