Remove Puppet::Util::Cacher usage from Puppet::Util::Settings
The path attribute was being unnecessarily cached. The value is a LoadedFile
instance, which already knows how to check whether it needs to be reloaded. The
act of reparsing was being triggered separately from the cacher mechanism.

The comment indicated this value was only being cached so it could be easily
cleared for tests, but it wasn't being cleared for tests. Thus, there is no
reason for this attribute to be cached, so remove it.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>