(#7184) Centralize "find action for face" into Puppet::Face
As part of moving to load actions first, and their associated face, when
invoked from the command line, it makes sense to push the logic for finding
the action and face down into the Puppet::Face implementation.

This means that we can change the logic there without needing to update the
public part of the CLI implementation, and that any further facades can use
the same, correct, logic to locate the action for the face.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>