Confine password disclosure acceptance test to hosts with required libraries
The useradd provider has the requirement that ruby-shadow[1] be
installed to be able to manage passwords.

On systems where we would use the useradd provider and this library
has not been installed we don't bother running the test, since we will
never be able to see the output we are testing.

[1] http://ttsky.net/ruby/

Signed-off-by: Jacob Helwig <jacob@puppetlabs.com>
Reviewed-by: Dominic Maraglia <dominic@puppetlabs.com>
Reviewed-by: Nick Lewis <nick@puppetlabs.com>