(#7290) Fail on unknown options.
Part of the "social contract" of Faces, Actions and Options is that the
metadata we collect is authoritative: it covers everything that is possible.

In the initial release we didn't enforce that around options.  If you passed
an unknown option in the hash, we just silently ignored it in validation and
made it available down in the action.

Now, instead, we enforce that rule.  If you pass an unknown option we raise an
error and complain; anything that gets to the action will be listed in the set
of inspectable options.

Cases that depended on this behaviour to pass arbitrary content in the hash
should be rewritten to move that content down a level: take a hash value for
one option, and use that for your free content.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>