(#7290) Update indirected Faces to avoid unknown options.
Now that we enforce that options must be declared, the model we exposed in the
Ruby API (but not the CLI facade) was that you could pass additional arguments
to the indirection method by passing them as unknown options doesn't work.

Instead, explicitly declare an option, `extra`, that accepts the final
argument to be passed direct to the indirection.  This makes things work
smoothly, as well as making it possible (once you can input a hash on the
command line) to invoke extra arguments from the facade too...

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>