(#8272) Use symbols instead of booleans for enabled property on Windows
Because the enable property of the service type uses :true and :false as its
valid values, rather than true and false, we need to return :true and :false
from our enabled? method. Otherwise, the property was being synced every time
it was enabled or disabled, regardless of whether it was actually in sync or
not.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>