(#8663) Update the run_mode spec test on Windows to match the code
When running as root, the default conf and var directories on Windows
are currently puppet/etc and puppet/var within the windows
directory. Updated the spec tests to match what the code does on
Windows.

Whether or not this is the correct behavior is something that will be
addressed in the future. See #8660.

Reviewed-by: Jacob Helwig <jacob@puppetlabs.com>