(#8663) Drive letters are not valid absolute paths on Windows
One of the file spec tests was attempting to use X: as a fully
qualified file path, and the file type was correctly rejecting it,
since it is a relative path. This commit changes the spec test to
expect the file type to raise an exception.

Reviewed-by: Jacob Helwig <jacob@puppetlabs.com>
(cherry picked from commit e0d3f11fc7b580aedec5350dc41a01417d6acd8d)