(#8690) Accept 'global' options in Puppet Faces
When we introduced verification of options, we forgot to handle the case that
global options from the Puppet settings system could be passed to the face.
This, in turn, means that the system would fail if you used any of those.

This remediates that, and now these work as expected.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>