(#8660) Fix destdir option on Windows
Specifying the --destdir option failed on windows because the install
script attempted to concatenate two absolute paths together. On Unix,
this is fine, but on Windows, it fails because the colon (part of the
drive specifier) is not a valid filename character. This commit adds a
method to join two paths, stripping off the drive specifier, if
present.

Also fixed a minor bug, which caused a deprecation warning to always
be printed when the --destdir option is omitted, which is the default
case.

Reviewed-by: Cameron Thomas <cameron@puppetlabs.com>
(cherry picked from commit f6006177a85766cfb3387db6c3d337bf97b52f17)